# Replace OptimisticTransactionDB with DB and WriteBatchWithIndex 