This commit replaces the OptimisticTransactionDB with the normal DB
and using WriteBatchWithIndex instead of `Transactions`.

This fixes https://github.com/restatedev/restate/issues/1428.

This PR is based on #1886.

What did not come true is that the performance of Restate would improve with this change. Even though, we should be doing less work when it comes to writing the `WriteBatchWithIndex` and the same when reading, I observed a slight regression in throughput with this change on my machine (might be related to variance). @AhmedSoliman it would be great if you could double check this with your benchmark setup. I cannot really explain why this change (which on the paper removes work) ends up reducing throughput.