@AhmedSoliman 2024-08-26
Before:
```console
❯ wrk -t 8 -c 200 --latency -d3m  -s ~/workspace/get.lua http://127.0.0.1:8080
Running 3m test @ http://127.0.0.1:8080
  8 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    34.80ms   33.82ms 409.56ms   88.16%
    Req/Sec     0.87k   228.43     1.72k    67.09%
  Latency Distribution
     50%   25.15ms
     75%   44.69ms
     90%   74.20ms
     99%  167.62ms
  1212234 requests in 3.00m, 322.55MB read
  Socket errors: connect 0, read 66, write 0, timeout 200
Requests/sec:   6730.73
Transfer/sec:      1.79MB
```
After:
```console
❯ wrk -t 8 -c 200 --latency -d3m  -s ~/workspace/get.lua http://127.0.0.1:8080
Running 3m test @ http://127.0.0.1:8080
  8 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    32.12ms   30.25ms 356.33ms   88.62%
    Req/Sec     0.93k   221.30     1.62k    68.34%
  Latency Distribution
     50%   24.01ms
     75%   40.23ms
     90%   65.47ms
     99%  155.78ms
  1333989 requests in 3.00m, 354.94MB read
Requests/sec:   7405.40
Transfer/sec:      1.97MB
```

I've ran the test several times, the difference is within margin of error, so we don't see clear performance gain out of this change, but the simplification is 100% worth doing it.