@tillrohrmann 2024-08-26
Thanks a lot for measuring the performance impact @AhmedSoliman and the review. Merging this PR after rebasing.